Extracted Vim script parts into 'vimscript' package
