+++ src/com/maddyhome/idea/vim/ex/handler/MapHandler.java
@@ -24,6 +24,7 @@ import com.intellij.openapi.util.text.StringUtil;
 import com.maddyhome.idea.vim.VimPlugin;
 import com.maddyhome.idea.vim.command.MappingMode;
 import com.maddyhome.idea.vim.ex.*;
+import com.maddyhome.idea.vim.ex.vimscript.VimScriptCommandHandler;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
