+++ src/com/maddyhome/idea/vim/ex/handler/SourceHandler.java
@@ -21,6 +21,8 @@ package com.maddyhome.idea.vim.ex.handler;
 import com.intellij.openapi.actionSystem.DataContext;
 import com.intellij.openapi.editor.Editor;
 import com.maddyhome.idea.vim.ex.*;
+import com.maddyhome.idea.vim.ex.vimscript.VimScriptCommandHandler;
+import com.maddyhome.idea.vim.ex.vimscript.VimScriptParser;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
