+++ src/com/maddyhome/idea/vim/ex/vimscript/VimScriptCommandHandler.java
@@ -16,8 +16,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.maddyhome.idea.vim.ex;
+package com.maddyhome.idea.vim.ex.vimscript;
 
+import com.maddyhome.idea.vim.ex.ExCommand;
+import com.maddyhome.idea.vim.ex.ExException;
 import org.jetbrains.annotations.NotNull;
 
 /**
